+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - New test: replace command
 - New test: append command
 - New test: prepend command
+- New test: touch command
 
 ## 0.0.1 - 2022-02-01
 - Start protocol information details